Changed VXLAN Bridge name and CentOS 7 Agent Setup.
Changed the bridge naming convention from brvx-xxxx to vxbr-xxxx.  This
is to work around an issue with libvirt that will not handle migration
of VMs when using the "brvx" prefix.

Additionally, change the utilites.py to always force start
cloudstack-agent when using CentOS 7.  The "isRunning" method was
returning True even when the agent was not running which caused the
agent not to start and the host failed to add.  I beleive its related to
an error code 3 being returned by systemctl status when a service is not
running.