Fixed for VM Password Reset Issue
When a password reset is requested by the user for a VM via the GUI the
changed is pushed to the vmpassword.json file as IP/password pairs.  It
is then sent to the password server which stored it in a separate cache
file and sends it to the client when requested.  The password is cleared
from the cache but not from the vmpassword.json file.  This cause the
password reset to be sent again to the password server when ever the VM
is powered on.

The change deletes the IP/password pair from the vmpassword.json file
after it hands it off to the password server.