Storage limits checks changes.
Change the check for primary storage limits for the attach volume to not
pass the size of the disk.  This was causing an issue where the create
updated the DB and then attach would add the disk size again after the
DB update.

Change the delete volumen to recalculate primary storage when the volume
state was in "Allocated".